requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-cleanup==7.0.0
Pillow==9.5.0
pycodestyle==2.10.0
python-dotenv==1.0.0
sqlparse==0.4.3
tomli==2.0.1
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
